YEEEEE

+++ src/Main.cpp
@@ -233,8 +233,8 @@ vector<string> extractPeers(string s){
     for (int i = 0; i < s.length(); i += 6){
         string peer = "";
         for (int j = 0; j < 4; j++){
-            cerr << (byte) s[i+j] << endl;
-            peer += to_string((unsigned int) s[i+j]);
+            cerr << (uint8_t) s[i+j] << endl;
+            peer += to_string((uint8_t) s[i+j]);
             if (j != 3)
                 peer += '.';
         }
